requirements.txt
PyQt5==5.15.6
mysql-connector-python==8.0.27
cryptography==3.4.7
pycryptodome==3.10.1
mysql~=0.0.3